% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{auc_stats}
\alias{auc_stats}
\title{auc_stats}
\usage{
auc_stats(
  data,
  response,
  case,
  control = NULL,
  predictor,
  best_method = "youden"
)
}
\arguments{
\item{data}{dataframe}

\item{response}{response variable}
}
\description{
get statistics of AUC
}
