% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binarize_cat}
\alias{binarize_cat}
\title{binarize categorical variable}
\usage{
binarize_cat(
  x,
  pos,
  neg = NULL,
  return.binary = F,
  label.pos = NULL,
  label.neg = NULL
)
}
\arguments{
\item{x}{vector}

\item{pos}{positive values}

\item{neg}{negative values, if NULL, use all other values except 'pos', the non-pos,
non-neg values in x will be converted to NA}

\item{return.binary}{return the binary 0(negative), 1(positive), default FALSE}

\item{label.pos}{label for positive values, default 'pos'}

\item{label.neg}{label for negative values, default 'neg'}
}
\description{
binarize categorical variable
}
