% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{binarize_data}
\alias{binarize_data}
\title{binarize categorical or numeric data}
\usage{
binarize_data(
  x,
  datatype = "auto",
  return_binary = F,
  label_pos = NULL,
  label_neg = NULL,
  cat_pos = NULL,
  cat_neg = NULL,
  num_cut_method = "median",
  response = NULL,
  response_pos = NULL,
  response_neg = NULL,
  auc_best_method = "youden",
  time = NULL,
  event = NULL,
  surv_minprop = 0.3
)
}
\arguments{
\item{x}{input vector}

\item{datatype}{"num" or "cat", default "auto", which will get by \code{\link{datatype_num_cat}(x)}}

\item{return_binary}{return the binary 0(negative), 1(positive), default FALSE}

\item{label_pos}{label for positive values, default 'pos'}

\item{label_neg}{label for negative values, default 'neg'}

\item{cat_pos}{if x is categorical, the positive group values}

\item{cat_neg}{if x is categorical, the negative group values}

\item{num_cut_method}{cut method for numeric variable}

\item{response}{response variable}

\item{response_pos}{positive values for response}

\item{response_neg}{negative values for response}

\item{auc_best_method}{best method using AUC analysis. Possible values include: "youden"(default), "closest.topleft"}

\item{time}{survival time}

\item{event}{survival event}

\item{surv_minprop}{optimal cutpoint based on survival, minimum proporal}
}
\value{
list containing 'data' and 'cutpoint', 'data' is the binary response,
'cutpoint' is NA if input isn't numeric variable
}
\description{
binarize categorical or numeric data
}
