% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binarize_num}
\alias{binarize_num}
\title{binarize numeric values}
\usage{
binarize_num(
  x,
  cutpoint,
  return.binary = F,
  label.pos = "high",
  label.neg = "low"
)
}
\arguments{
\item{x}{vector of numeric}

\item{cutpoint}{the cutpoint}

\item{label.pos}{label for 'positive' group, default 'high'}

\item{label.neg}{label for 'negative' group, default 'low'}

\item{response}{necessary if cut = "youden_ROC"}
}
\value{
binary value
}
\description{
binarize numeric values by predefined cutpoint
}
