% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clin_bmk_imvigor210}
\alias{clin_bmk_imvigor210}
\title{clinical and bimarker data of Imvigor210}
\format{
A data frame with 348 observations on the following 17965 variables:
\describe{
  \item{price}{price, in US dollars}
  \item{carat}{weight of the diamond, in carats}
  ...
}
}
\source{
\url{http://research-pub.gene.com/IMvigor210CoreBiologies/packageVersions/IMvigor210CoreBiologies_1.0.0.tar.gz}
}
\usage{
clin_bmk_imvigor210
}
\description{
A dataset containing the clinical and gene expression variabels
}
\keyword{datasets}
