% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{cutpoint}
\alias{cutpoint}
\title{cutpoint}
\usage{
cutpoint(
  x,
  method = "median",
  response = NULL,
  response_pos = NULL,
  response_neg = NULL,
  auc_best_method = "youden",
  time = NULL,
  event = NULL,
  surv_minprop = 0.3
)
}
\arguments{
\item{x}{vector}

\item{method}{median(default), roc, optimal_surv}

\item{response}{used if method='roc'}

\item{response_pos}{used if method='roc', positive response label}

\item{response_neg}{used if method='roc', negative response label, if NULL, use all other samples except response_pos}

\item{auc_best_method}{'youden' or 'closest.topleft', used if method='roc'}

\item{time}{used if method='optimal_surv_cut', survival time used for}

\item{surv_minprop}{used if method='optimal_surv_cut', min proportion in high and low group}
}
\value{
cutpoint
}
\description{
get cutpoint
}
