% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{datatype_num_cat}
\alias{datatype_num_cat}
\title{get datatype "num" or "cat"}
\usage{
datatype_num_cat(x)
}
\arguments{
\item{x}{vector}
}
\value{
"num" or "cat"
}
\description{
get datatype "num" or "cat"
}
