% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_4quadrant_response.R,
%   R/four_quadrants_response.R
\name{dm_4quadrant_response}
\alias{dm_4quadrant_response}
\title{Four quadrant analysis for response}
\usage{
dm_4quadrant_response(
  data,
  response,
  response_pos,
  response_neg = NULL,
  marker1,
  marker2,
  num_cut_method = "none",
  m1_cat_pos = NULL,
  m1_cat_neg = NULL,
  m2_cat_pos = NULL,
  m2_cat_neg = NULL,
  na_rm = T
)

dm_4quadrant_response(
  data,
  response,
  response_pos,
  response_neg = NULL,
  marker1,
  marker2,
  num_cut_method = "none",
  m1_cat_pos = NULL,
  m1_cat_neg = NULL,
  m2_cat_pos = NULL,
  m2_cat_neg = NULL,
  na_rm = T
)
}
\arguments{
\item{data}{data frame}

\item{response}{response variable}

\item{response_pos}{positive value(s) of response variable}

\item{response_neg}{negative value(s) of response variable, default NULL, i.e. all values except positive value(s)}

\item{marker1}{marker1 variable}

\item{marker2}{marker2 variable}

\item{num_cut_method}{marker cut method, possible values include 'none'(default), 'roc', 'median'}

\item{m1_cat_pos}{positive value(s) if marker1 is categorical variable}

\item{m1_cat_neg}{negative value(s) if marker1 is categorical variable}

\item{m2_cat_pos}{positive value(s) if marker2 is categorical variable}

\item{m2_cat_neg}{negative value(s) if marker2 is categorical variable}
}
\value{
list of 'data', 'param'

list of 'data', 'param'
}
\description{
evaluate the logistic regression of dual marker

evaluate the logistic regression of dual marker
}
