% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_combine_all.R
\name{dm_combineM_cox}
\alias{dm_combineM_cox}
\title{search M2 using Cox model}
\usage{
dm_combineM_cox(
  data,
  time,
  event,
  candidates,
  covariates = NULL,
  m.binarize,
  m.num.cut = "median",
  m.cat.pos = NULL,
  m.cat.neg = NULL,
  p.adjust.method = "BH"
)
}
\arguments{
\item{data}{dataframe}

\item{time}{survival time}

\item{candidates}{marker candidates}

\item{covariates}{confounding factor}

\item{m.binarize}{binarize marker1, default FALSE}

\item{m.num.cut}{cut method/values for numeric marker1}

\item{m.cat.pos}{positive value for categorical marker1}

\item{m.cat.neg}{negative value for categorical marker2}
}
\description{
search M2 to combine with M1 using Cox model
}
