% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_combine_all.R
\name{dm_combineM_logit}
\alias{dm_combineM_logit}
\title{search marker2 using logistic regression}
\usage{
dm_combineM_logit(
  data,
  response,
  response.pos,
  response.neg = NULL,
  candidates,
  covariates = NULL,
  m.binarize,
  m.num.cut = "median",
  m.cat.pos = NULL,
  m.cat.neg = NULL,
  auc = F,
  p.adjust.method = "BH"
)
}
\arguments{
\item{data}{dataframe}

\item{response}{response variable}

\item{response.pos}{positive value(s) for response variable}

\item{response.neg}{negative value(s) for response variable}

\item{covariates}{confounding factor}

\item{auc}{report AUC, default FALSE}

\item{marker1}{marker1}

\item{m2.candidates}{candidates marker2}

\item{m1.binarize}{binarize marker1, default FALSE}

\item{m2.binarize}{binarize marker2, default FALSE}

\item{m1.num.cut}{cut method/values for numeric marker1}

\item{m2.num.cut}{cut method/values for numeric marker2}

\item{m1.cat.pos}{positive value for categorical marker1}

\item{m1.cat.neg}{negative value for categorical marker1}

\item{m2.cat.pos}{positive value for categorical marker2}

\item{m2.cat.neg}{negative value for categorical marker2}
}
\description{
search marker2 to combine with marker1
}
