% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_model.R
\name{dm_cox}
\alias{dm_cox}
\title{dm_cox
evaluate the logistic regression of dual marker}
\usage{
dm_cox(
  data,
  time,
  event,
  marker1,
  marker2,
  binarization = F,
  num_cut_method = "none",
  m1_cat_pos = NULL,
  m1_cat_neg = NULL,
  m2_cat_pos = NULL,
  m2_cat_neg = NULL,
  na.rm = T
)
}
\arguments{
\item{num_cut_method}{marker cut method, [none, roc, median]}
}
\value{
dual marker logistic regression summary
}
\description{
dm_cox
evaluate the logistic regression of dual marker
}
