% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{dm_logit}
\alias{dm_logit}
\title{Logistic regression for dual marker}
\usage{
dm_logit(
  data,
  response,
  response_pos,
  response_neg = NULL,
  marker1,
  marker2,
  binarization = F,
  binary_regression = T,
  num_cut_method = "none",
  m1_cat_pos = NULL,
  m1_cat_neg = NULL,
  m2_cat_pos = NULL,
  m2_cat_neg = NULL,
  na_rm = T
)
}
\arguments{
\item{data}{data.frame}

\item{response}{response variables}

\item{response_pos}{positive values of response}

\item{response_neg}{negative values of response}

\item{marker1}{marker1 variable}

\item{marker2}{marker2 variable}

\item{num_cut_method}{marker cut method. Possible values can be 'none'(default) and 'roc', 'median'}

\item{m1_cat_pos}{positive value(s) if marker1 is categorical}

\item{m1_cat_neg}{negative value(s) if marker1 is categorical}

\item{m2_cat_pos}{positive value(s) if marker2 is categorical}

\item{m2_cat_neg}{negative value(s) if marker2 is categorical}

\item{na_rm}{remove NA, default TRUE}
}
\value{
summary of dual marker logistic regression
}
\description{
Evaluate the logistic regression of dual marker
}
