% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_pair.R
\name{dm_pair}
\alias{dm_pair}
\title{plot and statistics of dual marker pair}
\usage{
dm_pair(
  data,
  marker1,
  marker2,
  response = NULL,
  response.pos = NULL,
  response.neg = NULL,
  time = NULL,
  event = NULL,
  m1.num.cut = "median",
  m1.cat.pos = NULL,
  m1.cat.neg = NULL,
  m2.num.cut = "median",
  m2.cat.pos = NULL,
  m2.cat.neg = NULL,
  plot.only = F
)
}
\arguments{
\item{data}{data frame}

\item{marker1}{marker1}

\item{marker2}{marker2}

\item{response}{response variable}

\item{response.pos}{positive value(s) of response}

\item{response.neg}{negative value(s) of response, default NULL, all other values except 'response.pos'}

\item{time}{survival time}

\item{event}{survival event}

\item{m1.num.cut}{cut method/values for numeric marker1}

\item{m1.cat.pos}{positive value for categorical marker1}

\item{m1.cat.neg}{negative value for categorical marker1}

\item{m2.num.cut}{cut method/values for numeric marker2}

\item{m2.cat.pos}{positive value for categorical marker2}

\item{m2.cat.neg}{negative value for categorical marker2}

\item{plot.only}{no 'stats', default TRUE}
}
\value{
list of 'plot' and 'stats'.
}
\description{
the plot and statistics of dual markers, including the response and survival analysis
}
