% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_response_4quadrants.R
\name{dm_response_4quad}
\alias{dm_response_4quad}
\title{Four quadrant analysis for response}
\usage{
dm_response_4quad(
  data,
  response,
  response.pos,
  response.neg = NULL,
  marker1,
  marker2,
  m1.num.cut = "median",
  m1.cat.pos = NULL,
  m1.cat.neg = NULL,
  m2.num.cut = "median",
  m2.cat.pos = NULL,
  m2.cat.neg = NULL,
  na.rm = T
)
}
\arguments{
\item{data}{data frame}

\item{response}{response variable}

\item{response.pos}{positive value(s) of response variable}

\item{response.neg}{negative value(s) of response variable, default NULL, i.e. all values except positive value(s)}

\item{marker1}{marker1 variable}

\item{marker2}{marker2 variable}

\item{m1.num.cut}{cut method/value(s) if marker1 is numeric}

\item{m1.cat.pos}{positive value(s) if marker1 is categorical variable}

\item{m1.cat.neg}{negative value(s) if marker1 is categorical variable}

\item{m2.num.cut}{cut method/value(s) if marker2 is numeric}

\item{m2.cat.pos}{positive value(s) if marker2 is categorical variable}

\item{m2.cat.neg}{negative value(s) if marker2 is categorical variable}

\item{na.rm}{remove NA, default TRUE}
}
\value{
list of 'pos.n', 'total.n', 'stats', 'test', 'param'
}
\description{
evaluate the logistic regression of dual marker
}
