% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_response_boxplot.R
\name{dm_response_boxplot}
\alias{dm_response_boxplot}
\title{boxplot of responder and non-responder}
\usage{
dm_response_boxplot(
  data,
  response,
  response.pos,
  response.neg = NULL,
  marker1,
  marker2,
  m1.datatype = "auto",
  m1.cat.pos = NULL,
  m1.cat.neg = NULL,
  m2.datatype = "auto",
  m2.cat.pos = NULL,
  m2.cat.neg = NULL
)
}
\arguments{
\item{data}{dataframe}

\item{response}{response variable}

\item{response.pos}{positive value(s) for response, e.g. "CR_PR"}

\item{response.neg}{negative value(s) for response e.g. "SD_PD"}

\item{marker1}{marker1}

\item{marker2}{marker2}

\item{m1.datatype}{data type of marker1}

\item{m1.cat.pos}{positive value(s) for marker1 if marker1 is categorical, e.g. "MUT"}

\item{m1.cat.neg}{negative value(s) for marker1 if marker1 is categorical, e.g. "WT"}

\item{m2.datatype}{data type of marker2}

\item{m2.cat.pos}{positive value(s) for marker2 if marker2 is categorical, e.g. "MUT"}

\item{m2.cat.neg}{negative value(s) for marker2 if marker2 is categorical, e.g. "WT"}
}
\value{
panel of two ggplot object
}
\description{
boxplot of responder and non-responder
}
