% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roc_curve.R
\name{dm_roc}
\alias{dm_roc}
\title{plot ROC curve for dual markers}
\usage{
dm_roc(
  data,
  response,
  response_pos,
  response_neg = NULL,
  marker1,
  marker2,
  m1_cat_pos = NULL,
  m1_cat_neg = NULL,
  m2_cat_pos = NULL,
  m2_cat_neg = NULL,
  logistic_reg = TRUE,
  logistic_reg_int = FALSE
)
}
\arguments{
\item{data}{data.frame}

\item{response}{response}

\item{response_pos}{positive values of response}

\item{response_neg}{negative values of response, default NULL, i.e.g all values except response_pos}

\item{marker1}{marker1}

\item{marker2}{marker2}

\item{m1_cat_pos}{if m1 is categorical variable, the positve values}

\item{m1_cat_neg}{if m1 is categorical variable, the negative values}

\item{m2_cat_pos}{if m2 is categorical variable, the positve values}

\item{m2_cat_neg}{if m2 is categorical variable, the negative values}

\item{logistic_reg}{add logistic regression curve, no interaction, default TRUE}

\item{logistic_reg_int}{add logistic regression(with interaction) curve, default FALSE}
}
\value{
ROC curve, ggplot object
}
\description{
draw ROC curve for marker1, marker2 and fitted values by logistic regression w/ or w/o interaction between two markers
}
