% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_chart.R
\name{dm_scatter_chart}
\alias{dm_scatter_chart}
\title{Scatter-chart}
\usage{
dm_scatter_chart(
  data,
  response,
  response_pos,
  response_neg = NULL,
  marker1,
  marker2,
  m1_datatype = "auto",
  m2_datatype = "auto",
  num_cut_method = "median",
  m1_cat_pos = NULL,
  m1_cat_neg = NULL,
  m2_cat_pos = NULL,
  m2_cat_neg = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{response}{response/response variables}

\item{response_pos}{positive value(s) for response/response}

\item{response_neg}{negative value(s) for response/response}

\item{marker1}{marker1 variable}

\item{marker2}{marker2 variable}

\item{m1_datatype}{data type for marker1, 'num' or 'cat', default 'auto'}

\item{m2_datatype}{data type for marker1, 'num' or 'cat', default 'auto'}

\item{num_cut_method}{cut method if markers are numeric}

\item{m1_cat_pos}{positive value(s) if marker1 is categorical}

\item{m1_cat_neg}{negative value(s) if marker1 is categorical}

\item{m2_cat_pos}{positive value(s) if marker2 is categorical}

\item{m2_cat_neg}{negative value(s) if marker2 is categorical}
}
\description{
scatter-like chart to show relationship between two markers. Every sample is
presented as one dot, whose color indicates the response/response if 'response'
is assigned. If both markers are continuous variable, show scatter-plot. If
one marker is continuous, another one is categorical, show strip-plot. If
both are categorical, show jitter-plot.
}
