% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_searchM2.R
\name{dm_searchM2_cox}
\alias{dm_searchM2_cox}
\title{search M2 using Cox model}
\usage{
dm_searchM2_cox(
  data,
  time,
  event,
  marker1,
  m2.candidates,
  covariates = NULL,
  m1.binarize = F,
  m2.binarize = F,
  m1.num.cut = "median",
  m1.cat.pos = NULL,
  m1.cat.neg = NULL,
  m2.num.cut = "median",
  m2.cat.pos = NULL,
  m2.cat.neg = NULL,
  p.adjust.method = "BH"
)
}
\arguments{
\item{data}{dataframe}

\item{time}{survival time}

\item{event}{survival event}

\item{marker1}{marker1}

\item{m2.candidates}{marker2 candidates}

\item{covariates}{confounding factor}

\item{m1.binarize}{binarize marker1, default FALSE}

\item{m2.binarize}{binarize marker2, default FALSE}

\item{m1.num.cut}{cut method/values for numeric marker1}

\item{m1.cat.pos}{positive value for categorical marker1}

\item{m1.cat.neg}{negative value for categorical marker1}

\item{m2.num.cut}{cut method/values for numeric marker2}

\item{m2.cat.pos}{positive value for categorical marker2}

\item{m2.cat.neg}{negative value for categorical marker2}

\item{p.adjust.method}{see also p.adjust.methods}
}
\description{
search M2 to combine with M1 using Cox model
}
