% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualmarker_finder.R
\name{dm_searchM2_cox_plot}
\alias{dm_searchM2_cox_plot}
\title{plot of most significant marker2}
\usage{
dm_searchM2_cox_plot(res.searchM2, padj = F, max.n = 30, pval.threh = 0.01)
}
\description{
plot of most significant marker2
}
