% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_searchM2.R
\name{dm_searchM2_logit}
\alias{dm_searchM2_logit}
\title{search marker2 using logistic regression}
\usage{
dm_searchM2_logit(
  data,
  response,
  response.pos,
  response.neg = NULL,
  marker1,
  m2.candidates,
  confound.factor = NULL,
  m1.binarize = F,
  m2.binarize = F,
  m1.num.cut = "median",
  m1.cat.pos = NULL,
  m1.cat.neg = NULL,
  m2.num.cut = "median",
  m2.cat.pos = NULL,
  m2.cat.neg = NULL,
  na.rm = T,
  auc = F
)
}
\arguments{
\item{data}{dataframe}

\item{response}{response variable}

\item{response.pos}{positive value(s) for response variable}

\item{response.neg}{negative value(s) for response variable}

\item{marker1}{marker1}

\item{m2.candidates}{candidates marker2}

\item{confound.factor}{confounding factor}

\item{m1.binarize}{binarize marker1, default FALSE}

\item{m2.binarize}{binarize marker2, default FALSE}

\item{m1.num.cut}{cut method/values for numeric marker1}

\item{m1.cat.pos}{positive value for categorical marker1}

\item{m1.cat.neg}{negative value for categorical marker1}

\item{m2.num.cut}{cut method/values for numeric marker2}

\item{m2.cat.pos}{positive value for categorical marker2}

\item{m2.cat.neg}{negative value for categorical marker2}

\item{na.rm}{remove NA, default TRUE}

\item{auc}{report AUC, default FALSE}
}
\description{
search marker2 to combine with marker1
}
