% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_searchM2.R
\name{dm_searchM2_topPlot}
\alias{dm_searchM2_topPlot}
\title{plot search M2 result}
\usage{
dm_searchM2_topPlot(res_searchM2, top.n = 30, pval_threh = 0.01)
}
\arguments{
\item{res_searchM2}{result of search M2}

\item{pval_threh}{pvalue threhold}

\item{max_n}{max M2s to show}
}
\value{
an ggplot object
}
\description{
plot search M2 result
}
