% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_survival_KMplot.R
\name{dm_surv_plot}
\alias{dm_surv_plot}
\title{survival plot}
\usage{
dm_surv_plot(
  data,
  time,
  event,
  marker1,
  marker2,
  response = NULL,
  response.pos = NULL,
  response.neg = NULL,
  m1.num.cut = "median",
  m1.cat.pos = NULL,
  m1.cat.neg = NULL,
  m2.num.cut = "median",
  m2.cat.pos = NULL,
  m2.cat.neg = NULL,
  na.rm = T,
  km.pval = T,
  km.risk.table = T,
  ...
)
}
\arguments{
\item{data}{dataframe}

\item{time}{time}

\item{event}{event}

\item{marker1}{marker1}

\item{marker2}{marker2}

\item{response}{response}

\item{response.pos}{positive value(s) for response}

\item{response.neg}{negative value(s) for response}

\item{m1.num.cut}{cut method/value(s) if marker1 is numeric}

\item{m1.cat.pos}{positive value if marker1 is categorical}

\item{m1.cat.neg}{negative value if marker1 is categorical}

\item{m2.num.cut}{cut method/value(s) if marker2 is numeric}

\item{m2.cat.pos}{positive value if marker2 is categorical}

\item{m2.cat.neg}{negative value if marker2 is categorical}

\item{na.rm}{remove NA}

\item{km.pval}{show pval for KM-plot}

\item{km.risk.table}{show risk.table for KM-plot}

\item{...}{other parameter for ggsurvplot}
}
\description{
survival plot
}
