% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_survival_4quadrants.R
\name{dm_survival_4quad}
\alias{dm_survival_4quad}
\title{dual marker survival analysis of four quadrants}
\usage{
dm_survival_4quad(
  data,
  time,
  event,
  marker1,
  marker2,
  m1.datatype = "auto",
  m1.num.cut = "median",
  m1.cat.pos = NULL,
  m1.cat.neg = NULL,
  m2.datatype = "auto",
  m2.num.cut = "median",
  m2.cat.pos = NULL,
  m2.cat.neg = NULL
)
}
\arguments{
\item{data}{dataframe}

\item{time}{survival time}

\item{event}{survival event}

\item{marker1}{marker1}

\item{marker2}{marker2}

\item{m1.datatype}{data type of marker1}

\item{m1.num.cut}{cut method/value(s) if marker1 is numeric}

\item{m1.cat.pos}{positive values for marker1 if it is categorical}

\item{m1.cat.neg}{negative values for marker1 if it is categorical}

\item{m2.datatype}{data type of marker2}

\item{m2.num.cut}{cut method/value(s) if marker2 is numeric}

\item{m2.cat.pos}{positive values for marker2 if it is categorical}

\item{m2.cat.neg}{negative values for marker2 if it is categorical}
}
\description{
dual marker survival analysis of four quadrants
}
