% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.assert_colname}
\alias{.assert_colname}
\title{check the variable in colnames}
\usage{
.assert_colname(data, var)
}
\arguments{
\item{data}{dataframe}

\item{var}{variable names}
}
\description{
check the variable in colnames
}
