% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.collapse_df}
\alias{.collapse_df}
\title{collapse data.frame to one-row tibble}
\usage{
.collapse_df(df, id_cols, var_cols, sep = "_")
}
\arguments{
\item{df}{data frame}

\item{id_cols}{id cols, unique}

\item{var_cols}{var cols}

\item{sep}{separator between id and value}
}
\value{
one-row of tibble, each column is the
}
\description{
collapse data.frame to one-row tibble
}
