% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_4quadrant_response.R,
%   R/four_quadrants_response.R
\name{.dm_4quadrant_core}
\alias{.dm_4quadrant_core}
\title{four quadrant analysis for response(core)}
\usage{
.dm_4quadrant_core(data, response, marker1, marker2, na_rm = T)

.dm_4quadrant_core(data, response, marker1, marker2, na_rm = T)
}
\arguments{
\item{data}{data frame}

\item{response}{response variable, factor with level of 'neg' and 'pos'}

\item{marker1}{marker1 variable, factor with 2 levels and 1st level is negative}

\item{marker2}{marker2 variable, factor with 2 levels and 1st level is negative}
}
\description{
four quadrant analysis for response(core)

four quadrant analysis for response(core)
}
