% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cox_model.R
\name{.dm_cox_core}
\alias{.dm_cox_core}
\title{dm_cox_core
evaluate the logistic regression of dual marker}
\usage{
.dm_cox_core(data, time, event, marker1, marker2, na.rm = T)
}
\arguments{
\item{marker2}{}
}
\description{
dm_cox_core
evaluate the logistic regression of dual marker
}
