% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_chart.R
\name{.dm_jitter}
\alias{.dm_jitter}
\title{dm_jitter}
\usage{
.dm_jitter(
  data,
  marker1,
  marker2,
  response,
  response_pos,
  response_neg = NULL,
  m1_cat_pos,
  m1_cat_neg = NULL,
  m2_cat_pos,
  m2_cat_neg = NULL
)
}
\arguments{
\item{data}{data frame}

\item{marker1}{marker1 variable}

\item{marker2}{marker2 variable}

\item{response}{response variable}

\item{response_pos}{positive value(s) of response}

\item{response_neg}{negative value(s) of response}

\item{m1_cat_pos}{positive value(s) if marker1 is categorical}

\item{m1_cat_neg}{negative value(s) if marker1 is categorical}

\item{m2_cat_pos}{positive value(s) if marker2 is categorical}

\item{m2_cat_neg}{negative value(s) if marker2 is categorical}
}
\description{
jitter for marker1(categorical), marker2(categorical) and response
}
