% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{.dm_logit_core}
\alias{.dm_logit_core}
\title{dual marker analysis for logistic regression(core)}
\usage{
.dm_logit_core(
  data,
  response,
  marker1,
  marker2,
  binary_regression = T,
  na_rm = T
)
}
\arguments{
\item{data}{data frame}

\item{response}{factor, 2 level, 'neg' and 'pos'}
}
\description{
evaluate the logistic regression of dual marker
}
