% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_response_4quadrants.R
\name{.dm_reponse_4quad_core}
\alias{.dm_reponse_4quad_core}
\title{four quadrant analysis for response(core)}
\usage{
.dm_reponse_4quad_core(data, response, marker1, marker2, na.rm = T)
}
\arguments{
\item{data}{data frame}

\item{response}{response variable, factor with level of 'neg' and 'pos'}

\item{marker1}{marker1 variable, factor with 2 levels and 1st level is negative}

\item{marker2}{marker2 variable, factor with 2 levels and 1st level is negative}
}
\description{
four quadrant analysis for response(core)
}
