% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_chart.R
\name{.dm_scatterplot}
\alias{.dm_scatterplot}
\title{scatterplot}
\usage{
.dm_scatterplot(
  data,
  marker1,
  marker2,
  response,
  m1_cutpoint,
  m2_cutpoint,
  response_pos,
  response_neg = NULL
)
}
\arguments{
\item{data}{data frame}

\item{marker1}{marker1 variable}

\item{marker2}{marker2 variable}

\item{response}{response variable}

\item{m1_cutpoint}{cutpoint for marker1}

\item{m2_cutpoint}{cutpoint for marker2}

\item{response_pos}{positive response value(s)}

\item{response_neg}{negative response value(s)}
}
\description{
scatter plot with cutoff lines
}
