% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scatter_chart.R
\name{.dm_stripplot}
\alias{.dm_stripplot}
\title{dm_stripplot}
\usage{
.dm_stripplot(
  data,
  marker1,
  marker2,
  response,
  response_pos,
  response_neg = NULL,
  m1_cat_pos,
  m1_cat_neg = NULL,
  m2_cutpoint
)
}
\arguments{
\item{data}{data.frame}

\item{marker1}{marker1}

\item{marker2}{marker2}

\item{response}{response variables}

\item{response_pos}{positive value(s) for response}

\item{response_neg}{negative value(s) for response}

\item{m1_cat_pos}{positive value(s) for marker1 if marker1 is categorical}

\item{m1_cat_neg}{negative value(s) for marker1 if marker1 is categorical}

\item{m2_cutpoint}{cutpoint for marker2}
}
\description{
boxplot for marker1(categorical), marker2(number) and response
}
