% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_survival_4quadrants.R
\name{.dm_survival_4quad_core}
\alias{.dm_survival_4quad_core}
\title{dual marker survival analysis for four quadrant(core)}
\usage{
.dm_survival_4quad_core(data, time, event, marker1, marker2)
}
\arguments{
\item{data}{dataframe}

\item{time}{survival time}

\item{event}{survival event}

\item{marker1}{marker1}

\item{marker2}{marker2}
}
\value{
list of 'stats' and 'test'
}
\description{
dual marker survival analysis for four quadrant(core)
}
