% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_survival_KMplot.R
\name{.dm_survival_scatter_chart}
\alias{.dm_survival_scatter_chart}
\title{scatter plot for survival info}
\usage{
.dm_survival_scatter_chart(
  data,
  time,
  event,
  marker1,
  marker2,
  response = NULL,
  response.pos = NULL,
  response.neg = NULL,
  m1.num.cut = "median",
  m1.cat.pos = NULL,
  m1.cat.neg = NULL,
  m2.num.cut = "median",
  m2.cat.pos = NULL,
  m2.cat.neg = NULL
)
}
\arguments{
\item{data}{dataframe}

\item{time}{time}

\item{event}{event}

\item{marker1}{marker1}

\item{marker2}{marker2}

\item{response}{response}

\item{response.pos}{positive values for response}

\item{response.neg}{negative values for response}

\item{m1.num.cut}{cut method/values if marker1 is numeric}

\item{m1.cat.pos}{positive value(s) if marker1 is categorical}

\item{m1.cat.neg}{negative value(s) if marker1 is categorical}

\item{m2.num.cut}{cut method/values if marker2 is numeric}

\item{m2.cat.pos}{positive value(s) if marker2 is categorical}

\item{m2.cat.neg}{negative value(s) if marker2 is categorical}
}
\description{
scatter plot for survival info
}
