% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.label_quadrant}
\alias{.label_quadrant}
\title{label quadrant}
\usage{
.label_quadrant(m1, m2)
}
\arguments{
\item{m1}{binary factor, 1st level as negative, 2nd level as positive}

\item{m2}{binary factor, 1st level as negative, 2nd level as positive}
}
\value{
factor with 4 levels, R1, R2, R3, R4
}
\description{
label four quadrants R1, R2, R3, R4, by binary input of marker1 and marker2
}
\details{
marker1, marker2 are binary factors, and first level as negative, 2nd level as positive
for (m1, m2):
(pos, pos) => R1, first quadrant
(neg, pos) => R2, 2nd quadrant
(neg, neg) => R3, 3rd quadrant
(pos, neg) => R4, 4th quadrant
}
