% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_quadrants_response_plot.R
\name{.quadrant_donut_chart}
\alias{.quadrant_donut_chart}
\title{draw donut plot}
\usage{
.quadrant_donut_chart(x, n)
}
\arguments{
\item{x}{positive counts in 1st, 2nd, 3rd and 4th quadrant}

\item{n}{total counts in 1st, 2nd, 3rd and 4th quadrant}
}
\description{
draw donut plot
}
