% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_quadrants_response_plot.R
\name{.quadrant_interact_chart}
\alias{.quadrant_interact_chart}
\title{bubble-pie plot to show positive ratio in each quadrant}
\usage{
.quadrant_interact_chart(x, n)
}
\arguments{
\item{x}{positive counts in 1st, 2nd, 3rd and 4th quadrant}

\item{n}{total counts in 1st, 2nd, 3rd and 4th quadrant}
}
\description{
bubble-pie plot to show positive ratio in each quadrant
}
