% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_response_4quadrants_chart.R
\name{.quadrant_matrix_chart}
\alias{.quadrant_matrix_chart}
\title{Quadrant Matrix Chart}
\usage{
.quadrant_matrix_chart(x, n)
}
\arguments{
\item{x}{positive counts in 1st, 2nd, 3rd and 4th quadrant}

\item{n}{total counts in 1st, 2nd, 3rd and 4th quadrant}
}
\description{
2X2 matrix plot showing positive ratio in each quadrant
}
