% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_quadrants_response_plot.R
\name{.quadrant_prop_chart}
\alias{.quadrant_prop_chart}
\title{draw four quadrant proportional area chart}
\usage{
.quadrant_prop_chart(x, test_method = "fisher.test")
}
\arguments{
\item{x}{vector of 4 numbers, for 1st, 2nd, 3rd and 4th quadrant}

\item{col_quad}{color of quadrant}

\item{col_text}{color of text}
}
\value{
ggplot object
}
\description{
this function plot
}
