% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_response_4quadrants.R
\name{.quadrant_stats_response}
\alias{.quadrant_stats_response}
\title{statistic of response rate in each quadrant}
\usage{
.quadrant_stats_response(x, n, combined_quadrant = F)
}
\arguments{
\item{x}{positive number in each quadrant, length(x)}

\item{n}{total number in each quadrant}
}
\value{
dataframe with n.total, n.pos, n.neg, pct.pos, pos.lower95,
  pos.upper95, region, .m1.level, .m2.level
}
\description{
statistic of response rate in each quadrant
}
