% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_survival_4quadrants.R
\name{.quadrant_survival_stats}
\alias{.quadrant_survival_stats}
\title{quadrant stats survival}
\usage{
.quadrant_survival_stats(data, time, event, marker1, marker2)
}
\arguments{
\item{data}{dataframe}

\item{time}{survival time}

\item{event}{survival event}

\item{marker1}{marker1}

\item{marker2}{marker2}
}
\value{

}
\description{
statistics of survival time in each quadrant
}
