% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_survival_4quadrants.R
\name{.quadrant_survival_test}
\alias{.quadrant_survival_test}
\title{quadrant survival test}
\usage{
.quadrant_survival_test(data, time, event, marker1, marker2, na.rm = T)
}
\arguments{
\item{data}{dataframe}

\item{time}{survival time}

\item{event}{survival event}

\item{marker1}{marker1}

\item{marker2}{marker2}

\item{na.rm}{remove NA}
}
\description{
log-rank test to compare each quadrants
}
