% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_4quadrant_response.R,
%   R/four_quadrants_response.R
\name{.quadrant_test_response}
\alias{.quadrant_test_response}
\title{test of response rate in different quadrant}
\usage{
.quadrant_test_response(x, n, method = "fisher.test", ...)

.quadrant_test_response(x, n, method = "fisher.test", ...)
}
\description{
compare response rate between quadrants

compare response rate between quadrants
}
\details{
R1 vs R4, R2 vs R3, R1 vs R2, R3 vs R4, R1 vs R3, R2 vs R4,
R14(R1+R4) vs R24(R2+R4), R12(R1+R2) vs R34(R3 + R4)

R1 vs R4, R2 vs R3, R1 vs R2, R3 vs R4, R1 vs R3, R2 vs R4,
R14(R1+R4) vs R24(R2+R4), R12(R1+R2) vs R34(R3 + R4)
}
