% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_survival_cox.R
\name{.sm_cox_core}
\alias{.sm_cox_core}
\title{dm_cox_core}
\usage{
.sm_cox_core(data, time, event, marker, covariates = NULL)
}
\arguments{
\item{data}{dataframe}

\item{time}{time}

\item{event}{event}

\item{marker}{marker}

\item{covariates}{covariates}
}
\value{
stats
}
\description{
Evaluate the logistic regression of dual marker
}
