% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_response_logit.R
\name{.sm_logit_core}
\alias{.sm_logit_core}
\title{single marker logit regression(core)}
\usage{
.sm_logit_core(
  data,
  response,
  marker,
  confound.factor = NULL,
  na.rm = T,
  auc = F
)
}
\arguments{
\item{data}{data frame}

\item{response}{response variable}

\item{marker}{marker}

\item{confound.factor}{confound.factor}

\item{na.rm}{remove NA}
}
\value{
summary of logit model
}
\description{
single marker logit regression(core)
}
