% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_survival_cox.R
\name{.summary_cox}
\alias{.summary_cox}
\title{summary of cox model}
\usage{
.summary_cox(cox.model, cindex = T)
}
\arguments{
\item{cox.model}{cox model}
}
\value{
a list of 'coef','glance' and 'cox.zph'
}
\description{
summary of cox model
}
