% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_response_logistic.R
\name{.summary_logit}
\alias{.summary_logit}
\title{summarize logistic regression model}
\usage{
.summary_logit(logit.model, auc = F)
}
\arguments{
\item{logit.model}{logistic regression model}
}
\value{
list of 'coef' and 'summary'
}
\description{
summarize logistic regression model, including the coefficient, AIC, and ROC/AUC of fitted values
}
