% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_survival_4quadrants.R
\name{.surv_test_binary_var}
\alias{.surv_test_binary_var}
\title{Survival test for binary variables}
\usage{
.surv_test_binary_var(data, var, time, event)
}
\arguments{
\item{data}{dataframe}

\item{var}{variable}

\item{time}{survival time}

\item{event}{survival event}
}
\value{
log-rank test and cox test result
}
\description{
test using log-rank and cox-test for binary variable
}
