% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_survival_cox.R
\name{sm_cox}
\alias{sm_cox}
\title{dual maker cox analysis}
\usage{
sm_cox(
  data,
  time,
  event,
  marker,
  covariates = NULL,
  binarize = F,
  num.cut = "none",
  cat.pos = NULL,
  cat.neg = NULL
)
}
\arguments{
\item{data}{dataframe}

\item{time}{survival time}

\item{event}{survival event}

\item{marker}{marker}

\item{covariates}{confounding factors}

\item{binarize}{binarize}

\item{num.cut}{marker cut method, [none, roc, median, mean]}

\item{cat.pos}{positive value(s) if marker is categorical}

\item{cat.neg}{negative value(s) if marker is categorical}
}
\value{
summary of cox model
}
\description{
dual marker survival analysis using Cox regression
}
