% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_searchM.R
\name{sm_searchM_logit}
\alias{sm_searchM_logit}
\title{Search single marker with logit model}
\usage{
sm_searchM_logit(
  data,
  response,
  response.pos,
  response.neg = NULL,
  candidates,
  covariates = NULL,
  binarize = F,
  num.cut = "none",
  cat.pos = NULL,
  cat.neg = NULL,
  auc = F
)
}
\arguments{
\item{data}{data frame}

\item{response}{response variable}

\item{response.pos}{positive values for response}

\item{response.neg}{negative values for response}

\item{candidates}{candidate markers}

\item{covariates}{confounding factors}

\item{binarize}{to binarize marker, default FALSE}

\item{num.cut}{cut method/values for numeric variable}

\item{cat.pos}{positive values for marker if marker is categorical}

\item{cat.neg}{negative values for marker if marker is categorical}

\item{auc}{statistic AUC, default FALSE}
}
\value{
dataframe
}
\description{
Search single marker with logit model
}
