% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_searchM.R
\name{sm_searchM_topPlot}
\alias{sm_searchM_topPlot}
\title{plot sm_searchM}
\usage{
sm_searchM_topPlot(res.searchM, top.n = 20)
}
\arguments{
\item{res.searchM}{result of sm_searchM}

\item{top.n}{label top-n}
}
\description{
plot sm_searchM
}
