% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logistic_regression.R
\name{summary_logistic_reg}
\alias{summary_logistic_reg}
\title{summarize logistic regression model}
\usage{
summary_logistic_reg(logistic_model)
}
\arguments{
\item{logistic_model}{logistic regression model}
}
\value{
list of 'coef' and 'summary'
}
\description{
summarize logistic regression model, including the coefficient, AIC, and ROC/AUC of fitted values
}
