% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{binarize_data}
\alias{binarize_data}
\title{binarize categorical or numeric data}
\usage{
binarize_data(
  x,
  datatype = "auto",
  return.binary = F,
  num.cut = "median",
  cat.pos = NULL,
  cat.neg = NULL,
  label.pos = NULL,
  label.neg = NULL
)
}
\arguments{
\item{x}{input vector}

\item{datatype}{"num" or "cat", default "auto", which will get by \code{\link{datatype_num_cat}(x)}}

\item{return.binary}{return the binary 0(negative), 1(positive), default FALSE}

\item{num.cut}{cut method for numeric variable}

\item{cat.pos}{if x is categorical, the positive group values}

\item{cat.neg}{if x is categorical, the negative group values}

\item{label.pos}{label for positive values, default 'pos'}

\item{label.neg}{label for negative values, default 'neg'}
}
\value{
list containing 'data' and 'cutpoint', 'data' is the binary response,
'cutpoint' is NA if input isn't numeric variable
}
\description{
binarize categorical or numeric data
}
