% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{binarize_num}
\alias{binarize_num}
\title{binarize numeric values}
\usage{
binarize_num(
  x,
  cutpoint,
  return_binary = F,
  label_pos = "high",
  label_neg = "low"
)
}
\arguments{
\item{x}{vector of numeric}

\item{cutpoint}{the cutpoint}

\item{label_pos}{label for 'positive' group, default 'high'}

\item{label_neg}{label for 'negative' group, default 'low'}

\item{response}{necessary if cut = "youden_ROC"}
}
\value{
binary value
}
\description{
binarize numeric values by predefined cutpoint
}
