% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{clin_bmk_IMvigor210}
\alias{clin_bmk_IMvigor210}
\title{clinical and bimarker data of Imvigor210}
\format{A data frame with 348 observations on the following 17965 variables:
\describe{
  \item{gep_}{gene expressions}
  \item{gepscore_}{gene signature score}
  \item{mut_}{gene mutation}
  ...
}}
\source{
\url{http://research-pub.gene.com/IMvigor210CoreBiologies/packageVersions/IMvigor210CoreBiologies_1.0.0.tar.gz}
}
\usage{
clin_bmk_IMvigor210
}
\description{
A dataset containing the clinical and gene expression variables
}
\keyword{datasets}
