% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cutpoint}
\alias{cutpoint}
\title{cutpoint}
\usage{
cutpoint(x, method = "median")
}
\arguments{
\item{x}{vector}

\item{method}{median(default), mean, mean+sd, mean-sd, upper-tertile, lower-tertile}

\item{response}{used if method='roc'}

\item{response.pos}{used if method='roc', positive response label}

\item{response.neg}{used if method='roc', negative response label, if NULL, use all other samples except response.pos}

\item{auc_best_method}{'youden' or 'closest.topleft', used if method='roc'}
}
\value{
cutpoint
}
\description{
get cutpoint
}
