% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_survival_cox.R
\name{dm_cox}
\alias{dm_cox}
\title{dual maker cox analysis}
\usage{
dm_cox(
  data,
  time,
  event,
  marker1,
  marker2,
  confound.factor = NULL,
  m1.binarize,
  m2.binarize,
  m1.num.cut = "median",
  m1.cat.pos = NULL,
  m1.cat.neg = NULL,
  m2.num.cut = "median",
  m2.cat.pos = NULL,
  m2.cat.neg = NULL,
  na.rm = T
)
}
\arguments{
\item{data}{dataframe}

\item{time}{survival time}

\item{event}{survival event}

\item{marker1}{marker1}

\item{marker2}{marker2}

\item{confound.factor}{coufounding factors}

\item{m1.binarize}{binarize m1, default FALSE}

\item{m2.binarize}{binarize m2, default FALSE}

\item{m1.num.cut}{cut method/value(s) if marker is numeric}

\item{m1.cat.pos}{positive value(s) of marker1 if it is categorical}

\item{m1.cat.neg}{negative value(s) of marker1 if it is categorical}

\item{m2.num.cut}{cut method/value(s) if marker2 is numeric}

\item{m2.cat.pos}{positive value(s) of marker2 if it is categorical}

\item{m2.cat.neg}{negative value(s) of marker2 if it is categorical}

\item{na.rm}{remove NA}
}
\value{
summary of cox model
}
\description{
dual marker survival analysis using Cox regression
}
