% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMplot.R
\name{dm_KMplot}
\alias{dm_KMplot}
\title{KM plot for dual marker}
\usage{
dm_KMplot(
  data,
  time,
  event,
  marker1,
  marker2,
  num_cut_method = "median",
  response = NULL,
  response_pos = NULL,
  response_neg = NULL,
  m1_datatype = "auto",
  m1_cat_pos = NULL,
  m1_cat_neg = NULL,
  m1_label_pos = NULL,
  m1_label_neg = NULL,
  m2_datatype = "auto",
  m2_cat_pos = NULL,
  m2_cat_neg = NULL,
  m2_label_pos = NULL,
  m2_label_neg = NULL,
  na_rm = T
)
}
\arguments{
\item{data}{data frame}

\item{time}{survival time}

\item{event}{survival event}

\item{marker1}{marker1}

\item{marker2}{marker2}

\item{num_cut_method}{method to cut numeric variables}

\item{response}{response response}

\item{response_pos}{positive response response}

\item{response_neg}{negative response response}

\item{m1_datatype}{datatype of marker1}

\item{m1_cat_pos}{positive value(s) of marker1 if marker1 is categorical}

\item{m1_cat_neg}{negative value(s) of marker1 if marker1 is categorical}

\item{m1_label_pos}{label for positive response of marker1}

\item{m1_label_neg}{label for negative response of marker1}

\item{m2_datatype}{datatype of marker2}

\item{m2_cat_pos}{positive value(s) of marker2 if marker2 is categorical}

\item{m2_cat_neg}{positive value(s) of marker2 if marker2 is categorical}

\item{m2_label_pos}{label for positive response of marker2}

\item{m2_label_neg}{label for positive response of marker2}
}
\description{
KM plot for dual marker
}
