% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_pair.R
\name{dm_pair}
\alias{dm_pair}
\title{Visualization and statistics of dual marker pair}
\usage{
dm_pair(
  data,
  response = NULL,
  response.pos = NULL,
  response.neg = NULL,
  label.response.pos = "pos",
  label.response.neg = "neg",
  time = NULL,
  event = NULL,
  marker1,
  m1.datatype = "auto",
  m1.num.cut = "median",
  m1.cat.pos = NULL,
  m1.cat.neg = NULL,
  label.m1 = marker1,
  label.m1.pos = NULL,
  label.m1.neg = NULL,
  marker2,
  m2.datatype = "auto",
  m2.num.cut = "median",
  m2.cat.pos = NULL,
  m2.cat.neg = NULL,
  label.m2 = marker2,
  label.m2.pos = NULL,
  label.m2.neg = NULL,
  covariates = NULL,
  palette.4quadrant = "default",
  palette.other = "default",
  na.rm.response = F,
  na.rm.marker = F,
  output = c("plot", "stat")
)
}
\arguments{
\item{data}{data frame}

\item{response}{response variable}

\item{response.pos}{positive value(s) of response}

\item{response.neg}{negative value(s) of response, default NULL, i.e. all other values
except 'response.pos'}

\item{time}{survival time}

\item{event}{survival event}

\item{marker1}{marker1 variable}

\item{m1.num.cut}{cut method or value for numeric marker1, default "median", can be specific
value or 'median', 'mean', 'mean+sd', 'mean-sd', 'upper-tertile', 'lower-tertile'}

\item{m1.cat.pos}{positive value(s) if marker1 is categorical}

\item{m1.cat.neg}{negative value(s) if marker1 is categorical}

\item{label.m1}{label of marker1, default using marker1 variable}

\item{label.m1.pos}{label of positive value(s) for marker1}

\item{label.m1.neg}{label of negative value(s) for marker1}

\item{marker2}{marker2 variable}

\item{m2.num.cut}{cut method or value for numeric marker2, default "median"}

\item{m2.cat.pos}{positive value(s) if marker2 is categorical}

\item{m2.cat.neg}{negative value(s) if marker2 is categorical}

\item{label.m2}{label of marker2, default using marker2 variable}

\item{label.m2.pos}{label of positive value(s) for marker2}

\item{label.m2.neg}{label of negative value(s) for marker2}

\item{covariates}{covariates}

\item{palette.4quadrant}{the color palette to be used for coloring or filling four quadrants.
same with palette.other}

\item{palette.other}{the color palette to be used for coloring or filling by.
same as \code{\link[ggpubr]{get_palette}}
Allowed values include "grey" for grey color palettes; brewer palettes e.g. "RdBu", "Blues", ...; or custom
 color palette e.g. c("blue", "red"); and scientific journal palettes from ggsci R package,
 e.g.: "npg", "aaas", "lancet", "jco", "ucscgb", "uchicago", "simpsons" and "rickandmorty".}

\item{output}{output 'stat' and/or 'plot', default c("stat", "plot"), i.e both stat and plot}
}
\value{
list of 'response.plot', 'response.stats', 'survival.plot' and 'survival.stats'
}
\description{
Plot and statistics of dual markers that reveal the association between two markers,
response and survival. It takes the marker1, marker2, response or survival as input,
and returns plots (response.plot, survival.plot) and statistic results (response.stats,
survival.stats) from logistic regression for binary outcome(response) and Cox regression
for time-to-event outcome(survival).
}
\examples{

 ### example1: two biomarkers of continuous variables
 res.pair <- dm_pair(
  data = clin_bmk_IMvigor210,
  response = "binaryResponse",
  response.pos = "CR/PR",
  response.neg = "SD/PD",
  marker1 = "TMB",
  m1.num.cut = "median",
  marker2 = "gepscore_TGFb.19gene",
  m2.num.cut = "median"
 )
 res.pair$response.plot
 res.pair$response.stats

 ### example2: two biomarkers of continuous variables(custom cutoff)
 m1.cutoff <- quantile(clin_bmk_IMvigor210$TMB, 2/3, na.rm=T)
 m2.cutoff <- quantile(clin_bmk_IMvigor210$gepscore_TGFb.19gene, 1/3, na.rm=T)
 res.pair <- dm_pair(
  data = clin_bmk_IMvigor210,
  response = "binaryResponse",
  response.pos = "CR/PR",
  response.neg = "SD/PD",
  marker1 = "TMB",
  m1.num.cut = m1.cutoff,
  marker2 = "gepscore_TGFb.19gene",
  m2.num.cut = m2.cutoff
 )
 res.pair$response.plot
 res.pair$response.stats

 ### example3: continuous variable + categorical variable
 res.pair <- dm_pair(
    data = clin_bmk_IMvigor210,
    # survival info
    time = "os", event = "censOS",
    # marker1
    marker1 = "mut_ARID1A",
    m1.cat.pos = "YES", m1.cat.neg = "NO",
    # marker2
    marker2 = "gep_CXCL13",
    m2.num.cut = "median"
 )
 res.pair$survival.plot
 res.pair$survival.stats

 ### example4: two biomarkers of categorical variable
 res.pair <- dm_pair(
    data = clin_bmk_IMvigor210,
    # response info
    response = "binaryResponse",
    response.pos = "CR/PR",
    response.neg = "SD/PD",
    # survival info
    time = "os",
    event = "censOS",
    # marker1
    marker1 = "mut_ARID1A",
    m1.cat.pos = "YES",
    m1.cat.neg = "NO",
    label.m1.pos = "MUT",
    label.m1.neg ="WT",
    # marker2
    marker2 = "mut_TP53",
    m2.cat.pos = "YES",
    m2.cat.neg = "NO",
    label.m2.pos = "MUT",
    label.m2.neg = "WT",
    na.rm.response = F,
    na.rm.marker = F
 )
 res.pair$response.plot
 res.pair$response.stats
 res.pair$survival.plot
 res.pair$survival.stats

 ### example5: two biomarkers of continuous variables with covariates
 res.pair <- dm_pair(
  data = clin_bmk_IMvigor210,
  response = "binaryResponse",
  response.pos = "CR/PR",
  response.neg = "SD/PD",
  marker1 = "TMB",
  m1.num.cut = "median",
  marker2 = "gepscore_TGFb.19gene",
  m2.num.cut = "median",
  covariates = "Immune.phenotype"
 )
 res.pair$response.stats

 ### example6: adjust the palette, labels
 res.pair <- dm_pair(
  data = clin_bmk_IMvigor210,
  response = "binaryResponse",
  response.pos = "CR/PR",
  response.neg = "SD/PD",
  label.response.pos = "R",
  label.response.neg = "NR",
  marker1 = "TMB",
  m1.num.cut = "median",
  label.m1.pos = "TMB_hi",
  label.m1.neg = "TMB_lo",
  marker2 = "gepscore_TGFb.19gene",
  m2.num.cut = "median",
  label.m2 = "TGFbeta",
  label.m2.pos = "TGFb_hi",
  label.m2.neg = "TGFb_lo",
  palette.4quadrant = "jco",
  palette.other = "jco"
 )
 res.pair$response.plot

}
