% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_pair.R
\name{dm_pair}
\alias{dm_pair}
\title{Plot and statistics of dual marker pair}
\usage{
dm_pair(
  data,
  marker1,
  marker2,
  response = NULL,
  response.pos = NULL,
  response.neg = NULL,
  time = NULL,
  event = NULL,
  m1.num.cut = "median",
  m1.cat.pos = NULL,
  m1.cat.neg = NULL,
  m2.num.cut = "median",
  m2.cat.pos = NULL,
  m2.cat.neg = NULL,
  plot.only = F
)
}
\arguments{
\item{data}{data frame}

\item{marker1}{marker1 variable}

\item{marker2}{marker2 variable}

\item{response}{response variable}

\item{response.pos}{positive value(s) of response}

\item{response.neg}{negative value(s) of response, default NULL, all other values except 'response.pos'}

\item{time}{survival time}

\item{event}{survival event}

\item{m1.num.cut}{cut method or value for numeric marker1, default "median"}

\item{m1.cat.pos}{positive value(s) if marker1 is categorical}

\item{m1.cat.neg}{negative value(s) if marker1 is categorical}

\item{m2.num.cut}{cut method or value for numeric marker2, default "median"}

\item{m2.cat.pos}{positive value(s) if marker2 is categorical}

\item{m2.cat.neg}{negative value(s) if marker2 is categorical}

\item{plot.only}{plot only, no 'response.stats' or 'survival.stats', default FALSE}
}
\value{
list of 'response.plot', 'response.stats', 'survival.plot' and 'survival.stats'
}
\description{
Plot and statistics of dual markers that reveal the association between two markers, response and survival
}
\examples{

 ### example1: two biomarkers of continuous variables
 res.pair <- dm_pair(
  data = clin_bmk_IMvigor210,
  response = "binaryResponse",
  response.pos = "CR/PR", response.neg = "SD/PD",
  marker1 = "TMB",
  m1.num.cut = "median",
  marker2 = "gepscore_TGFb.19gene",
  m2.num.cut = "median")
 res.pair$response.plot
 res.pair$response.stats
 ### example2: two biomarkers of continuous variables(custom cutoff)
 m1.cutoff <- quantile(clin_bmk_IMvigor210$TMB, 2/3, na.rm=T)
 m2.cutoff <- quantile(clin_bmk_IMvigor210$gepscore_TGFb.19gene, 1/3, na.rm=T)
 res.pair <- dm_pair(
  data = clin_bmk_IMvigor210,
  response = "binaryResponse",
  response.pos = "CR/PR", response.neg = "SD/PD",
  marker1 = "TMB",
  m1.num.cut = m1.cutoff,
  marker2 = "gepscore_TGFb.19gene",
  m2.num.cut = m2.cutoff)
 res.pair$response.plot
 res.pair$response.stats
 res.pair$survival.plot
 res.pair$survival.stats

 ### example3: continuous variable + categorical variable
 res.pair <- dm_pair(
    data = clin_bmk_IMvigor210,
    # survival info
    time = "os", event = "censOS",
    # marker1
    marker1 = "mut_ARID1A",
    m1.cat.pos = "YES", m1.cat.neg = "NO",
    # marker2
    marker2 = "gep_CXCL13",
    m2.num.cut = "median")
 res.pair$survival.plot
 res.pair$survival.stats

 ### example4: two biomarkers of categorical variable
 res.pair <- dm_pair(
    data = clin_bmk_IMvigor210,
    # response info
    response = "binaryResponse",
    response.pos = "CR/PR", response.neg = "SD/PD",
    # survival info
    time = "os", event = "censOS",
    # marker1
    marker1 = "mut_ARID1A",
    m1.cat.pos = "YES", m1.cat.neg = "NO",
    # marker2
    marker2 = "Immune.phenotype",
    m2.cat.pos = "inflamed", m2.cat.neg = c("desert", "excluded"))
 res.pair$response.plot
 res.pair$response.stats
 res.pair$survival.plot
 res.pair$survival.stats

}
