% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_in_one.R
\name{dm_pair}
\alias{dm_pair}
\title{dual biomarker plot and statistics}
\usage{
dm_pair(
  data,
  response,
  response_pos,
  response_neg = NULL,
  marker1,
  marker2,
  time,
  event,
  num_cut_method = "median",
  m1_cat_pos = NULL,
  m1_cat_neg = NULL,
  m2_cat_pos = NULL,
  m2_cat_neg = NULL
)
}
\arguments{
\item{data}{data frame}

\item{response}{response variable}

\item{response_pos}{positive value(s) of response}

\item{response_neg}{negative vlaue(s) of response}

\item{marker1}{marker1}

\item{marker2}{marer2}

\item{time}{survival time}

\item{event}{survival event}

\item{num_cut_method}{cut method for numeric variable}

\item{m1_cat_pos}{positive value(s) if marker1 is categorical}

\item{m1_cat_neg}{negative value(s) if marker1 is categorical}

\item{m2_cat_pos}{positive value(s) if marker2 is categorical}

\item{m2_cat_neg}{negative value(s) if marker2 is categorical}
}
\description{
dual biomarker plot and statistics
}
