% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_response_scatter_chart.R
\name{dm_scatter_chart}
\alias{dm_scatter_chart}
\title{Scatter-chart}
\usage{
dm_scatter_chart(
  data,
  response,
  response.pos,
  response.neg = NULL,
  marker1,
  marker2,
  m1.num.cut = "median",
  m1.cat.pos = NULL,
  m1.cat.neg = NULL,
  m2.num.cut = "median",
  m2.cat.pos = NULL,
  m2.cat.neg = NULL,
  size = NULL,
  alpha = NULL,
  shape = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{response}{response/response variables}

\item{response.pos}{positive value(s) for response/response}

\item{response.neg}{negative value(s) for response/response}

\item{marker1}{marker1 variable}

\item{marker2}{marker2 variable}

\item{m1.num.cut}{cut method/value(s) if marker1 is numeric}

\item{m1.cat.pos}{positive value(s) if marker1 is categorical}

\item{m1.cat.neg}{negative value(s) if marker1 is categorical}

\item{m2.num.cut}{cut method/value(s) if marker2 is numeric}

\item{m2.cat.pos}{positive value(s) if marker2 is categorical}

\item{m2.cat.neg}{negative value(s) if marker2 is categorical}
}
\description{
scatter-like chart to show relationship between two markers. Every sample is
presented as one dot, whose color indicates the response/response if 'response'
is assigned. If both markers are continuous variable, show scatter-plot. If
one marker is continuous, another one is categorical, show strip-plot. If
both are categorical, show jitter-plot.
}
