% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualmarker_finder.R
\name{dm_searchM2_cox}
\alias{dm_searchM2_cox}
\title{find marker2 to combine with M1 for survival prediction}
\usage{
dm_searchM2_cox(
  data,
  time,
  event,
  marker1,
  targets,
  binarization = F,
  num_cut_method = "none",
  m1_cat_pos = NULL,
  m1_cat_neg = NULL,
  na.rm = T
)
}
\description{
find marker2 to combine with M1 for survival prediction
}
