% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dualmarker_finder.R
\name{dm_searchM2_logit}
\alias{dm_searchM2_logit}
\title{dm_searchM2_logit
search marker2 to combine with marker1}
\usage{
dm_searchM2_logit(
  data,
  response,
  response_pos,
  response_neg = NULL,
  marker1,
  targets,
  binarization = F,
  num_cut_method = "none",
  m1_cat_pos = NULL,
  m1_cat_neg = NULL,
  na.rm = T
)
}
\description{
dm_searchM2_logit
search marker2 to combine with marker1
}
