% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_searchM2.R
\name{dm_searchM2_topPlot}
\alias{dm_searchM2_topPlot}
\title{plot search M2 result}
\usage{
dm_searchM2_topPlot(
  res.searchM2,
  top.n = 30,
  show.padj = F,
  line.sig = c(0.01, 0.05)
)
}
\arguments{
\item{res.searchM2}{result of search M2}

\item{top.n}{top-N}
}
\value{
an ggplot object
}
\description{
plot search M2 result
}
