% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_response_scatter_chart.R
\name{.dm_jitter}
\alias{.dm_jitter}
\title{dm_jitter}
\usage{
.dm_jitter(
  data,
  marker1,
  marker2,
  response,
  response.pos,
  response.neg = NULL,
  m1.cat.pos,
  m1.cat.neg = NULL,
  m2.cat.pos,
  m2.cat.neg = NULL,
  size = NULL,
  shape = NULL,
  alpha = NULL
)
}
\arguments{
\item{data}{data frame}

\item{marker1}{marker1 variable}

\item{marker2}{marker2 variable}

\item{response}{response variable}

\item{response.pos}{positive value(s) of response}

\item{response.neg}{negative value(s) of response}

\item{m1.cat.pos}{positive value(s) if marker1 is categorical}

\item{m1.cat.neg}{negative value(s) if marker1 is categorical}

\item{m2.cat.pos}{positive value(s) if marker2 is categorical}

\item{m2.cat.neg}{negative value(s) if marker2 is categorical}
}
\description{
jitter for marker1(categorical), marker2(categorical) and response
}
