% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KMplot.R
\name{.dm_KMplot_core}
\alias{.dm_KMplot_core}
\title{KMplot of dual markers}
\usage{
.dm_KMplot_core(
  data,
  time,
  event,
  marker1,
  marker2,
  label_m1 = marker1,
  label_m2 = marker2,
  na_rm = T
)
}
\arguments{
\item{data}{data frame}

\item{time}{survival time}

\item{event}{survival event}

\item{marker1}{categorical variables with 2 levels}

\item{marker2}{categorical variables with 2 levels}

\item{label_m1}{label for marker1}

\item{label_m2}{label for marker2}

\item{na_rm}{remove NA}
}
\value{
list of ggplot object, for marker1, marker2 and dualmarker
}
\description{
KMplot of dual markers, for each single markers and dual markers
}
