% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_response_logistic.R
\name{.dm_logit_core}
\alias{.dm_logit_core}
\title{dual marker analysis for logistic regression(core)}
\usage{
.dm_logit_core(data, response, marker1, marker2, covariates = NULL, auc = T)
}
\arguments{
\item{data}{data frame}

\item{response}{factor, 2 level, 'neg' and 'pos'}

\item{marker1}{marker1}

\item{marker2}{marker2}
}
\description{
evaluate the logistic regression of dual marker
}
