% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_response_scatter_chart.R
\name{.dm_scatterplot}
\alias{.dm_scatterplot}
\title{scatterplot}
\usage{
.dm_scatterplot(
  data,
  marker1,
  marker2,
  response,
  response.pos,
  response.neg = NULL,
  m1.cutpoint,
  m2.cutpoint,
  size = NULL,
  alpha = NULL,
  shape = NULL,
  label.m1 = marker1,
  label.m2 = marker2,
  label.response.pos = "pos",
  label.response.neg = "neg",
  palette = "default",
  na.rm.response = T
)
}
\arguments{
\item{data}{data frame}

\item{marker1}{marker1 variable}

\item{marker2}{marker2 variable}

\item{response}{response variable}

\item{response.pos}{positive response value(s)}

\item{response.neg}{negative response value(s)}

\item{m1.cutpoint}{cutpoint for marker1}

\item{m2.cutpoint}{cutpoint for marker2}
}
\description{
scatter plot with cutoff lines
}
