% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_response_scatter_chart.R
\name{.dm_stripplot}
\alias{.dm_stripplot}
\title{dm_stripplot}
\usage{
.dm_stripplot(
  data,
  marker1,
  marker2,
  response,
  response.pos,
  response.neg = NULL,
  m1.cat.pos,
  m1.cat.neg = NULL,
  m2.cutpoint,
  shape = NULL,
  size = NULL,
  alpha = NULL
)
}
\arguments{
\item{data}{data.frame}

\item{marker1}{marker1}

\item{marker2}{marker2}

\item{response}{response variables}

\item{response.pos}{positive value(s) for response}

\item{response.neg}{negative value(s) for response}

\item{m1.cat.pos}{positive value(s) for marker1 if marker1 is categorical}

\item{m1.cat.neg}{negative value(s) for marker1 if marker1 is categorical}

\item{m2.cutpoint}{cutpoint for marker2}
}
\description{
boxplot for marker1(categorical), marker2(number) and response
}
