% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_response_4quadrants_chart.R
\name{.quadrant_donut_chart}
\alias{.quadrant_donut_chart}
\title{Quadrant doughnut chart}
\usage{
.quadrant_donut_chart(x, n)
}
\arguments{
\item{x}{positive counts in 1st, 2nd, 3rd and 4th quadrant}

\item{n}{total counts in 1st, 2nd, 3rd and 4th quadrant}
}
\description{
doughnut chart to show response rate in each quadrant
}
\details{
The width of doughnut reflects the relative sample size in the quadrant, and
red proportion reflects the response rate
}
