% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_response_4quadrants_chart.R
\name{.quadrant_interact_chart}
\alias{.quadrant_interact_chart}
\title{Quadrant Interaction Chart}
\usage{
.quadrant_interact_chart(x, n)
}
\arguments{
\item{x}{positive counts in 1st, 2nd, 3rd and 4th quadrant}

\item{n}{total counts in 1st, 2nd, 3rd and 4th quadrant}
}
\description{
line chart to show the response rate in each quadrant
}
