% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/four_quadrants_response_plot.R
\name{.quadrant_matrix_chart}
\alias{.quadrant_matrix_chart}
\title{show positive ratio in each quadrant using matrix}
\usage{
.quadrant_matrix_chart(x, n)
}
\arguments{
\item{x}{positive counts in 1st, 2nd, 3rd and 4th quadrant}

\item{n}{total counts in 1st, 2nd, 3rd and 4th quadrant}
}
\description{
show positive ratio in each quadrant using matrix
}
