% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dm_response_4quadrants_chart.R
\name{.quadrant_prop_chart}
\alias{.quadrant_prop_chart}
\title{Four quadrant proportional area chart}
\usage{
.quadrant_prop_chart(x, test_method = "fisher.test")
}
\arguments{
\item{x}{vector of 4 numbers, for sample numbers of 1st, 2nd, 3rd and 4th quadrant}
}
\value{
ggplot object
}
\description{
Four quadrant proportional area chart
}
