% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_response_logit.R
\name{.sm_logit_core}
\alias{.sm_logit_core}
\title{single marker logit regression(core)}
\usage{
.sm_logit_core(data, response, marker, covariates = NULL, auc = F)
}
\arguments{
\item{data}{data frame}

\item{response}{response variable}

\item{marker}{marker}

\item{covariates}{covariates}
}
\value{
summary of logit model
}
\description{
single marker logit regression(core)
}
