% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_response_logit.R
\name{sm_logit}
\alias{sm_logit}
\title{Single marker logit regression}
\usage{
sm_logit(
  data,
  response,
  response.pos,
  response.neg = NULL,
  marker,
  confound.factor = NULL,
  binarize = F,
  num.cut = "none",
  cat.pos = NULL,
  cat.neg = NULL,
  na.rm = T,
  auc = F
)
}
\arguments{
\item{data}{data frame}

\item{response}{response}

\item{response.pos}{positive values for response}

\item{response.neg}{negative values for response}

\item{marker}{marker}

\item{confound.factor}{confonding factor}

\item{binarize}{binarize for marker, default FALSE}

\item{num.cut}{cut method/value for numeric marker}

\item{cat.pos}{positive values for marker if marker is categorical}

\item{cat.neg}{negative values for marker if marker is categorical}

\item{na.rm}{remove NA}
}
\value{

}
\description{
Single marker logit regression
}
