% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sm_searchM.R
\name{sm_searchM_cox}
\alias{sm_searchM_cox}
\title{search marker using Cox model}
\usage{
sm_searchM_cox(
  data,
  time,
  event,
  candidates,
  confound.factor = NULL,
  binarize = F,
  num.cut = "none",
  cat.pos = NULL,
  cat.neg = NULL,
  na.rm = T
)
}
\arguments{
\item{data}{data frame}

\item{time}{survival time}

\item{event}{survival event}

\item{candidates}{candidates markers}

\item{confound.factor}{confounding factors}

\item{binarize}{to binarize the data}

\item{num.cut}{cut method/value for numeric variable}

\item{cat.pos}{positive value(s)}

\item{cat.neg}{negative value(s)}

\item{na.rm}{remove na, default TRUE}
}
\description{
search marker using Cox model
}
